Fix warnings found by clang.

+++ gridnav/main_gridnav.cc
@@ -12,7 +12,7 @@ int main(int argc, const char *argv[]) {
 
 	GridNav d(&map, x0, y0, xg, yg);
 	Result<GridNav> res = search<GridNav>(d, argc, argv);
-	dfpair(stdout, "final sol cost", "%u", d.pathcost(res.ops));
+	dfpair(stdout, "final sol cost", "%g", (double) d.pathcost(res.ops));
 
 	return 0;
 }
